Deduplicate fileChange location extraction
